feat(ui): disable search functionality for small select components
