+++ web/ui/react-app/src/components/generic/form-select-creatable.tsx
@@ -18,7 +18,7 @@ import { OptionType } from 'types/util';
 import { Position } from 'types/config';
 import { TooltipWithAriaProps } from './tooltip';
 import cx from 'classnames';
-import { formPadding } from './util';
+import { formPadding } from './form-shared';
 import { useError } from 'hooks/errors';
 
 type Props = {
@@ -237,6 +237,7 @@ const FormSelectCreatable: FC<FormSelectCreatableProps> = ({
 							}
 							components={customComponents}
 							styles={customStyles}
+							menuShouldScrollIntoView
 						/>
 					)}
 					rules={{
